backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-cors-headers==3.9.0
django-extensions==3.2.3
djangorestframework==3.15.1
filelock==3.13.1
gunicorn==20.1.0
idna==3.7
pillow==10.3.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
requests==2.32.3
setuptools==70.0.0
sqlparse==0.5.0
urllib3==2.2.1
virtualenv==20.25.0
